added creation of BTree using util and cache
util grabs the filename and returns 1 long string which is then broken up by the user's seqLen and these are converted into longs and then sequences.  if the sequence is a duplicate then it exists in the cache and its freq is incremented, else it is added to the cache and ultimately into the btree.  don't know how to dump the tree into a binary file yet